Flatbuffer Struct Padding issue (#62)
- Resolves #52 
- Upgrades flatbuffers (note user needs to upgrade flatc)


Note this is a breaking change for current contract users as now with
the forces 8 byte alignment account is now 40 bytes not 33. This means,
its aligned so its faster, but it means that once deployed, prover nodes
must expect 40 bytes or there will be errors.
At this stage Im certain we can communicate the change, and have no
breakage as additional callback accounts greater than the first one will
have been wrong. SInce we are pre 1.0 I think this is fine. But I see
that we could use this opportunity to test out the versioning process.
for the flat buffer structs.